requirements.txt
Django>=5.0,<5.2
gunicorn>=21.2,<22
psycopg[binary]>=3.1,<3.2
whitenoise>=6.6,<7
django-allauth>=0.63,<0.64
Pillow>=10.3,<11

requests>=2.31,<3
requests-oauthlib>=1.3,<2

PyJWT[crypto]>=2.8,<3
mailchimp-marketing==3.0.80
